Reverted to observing window dataloader
